// src/app/models/encounter/orm-chart-procedures.ts
export class ORMChartProcedures {
    chart_procedures_id: String;
    chart_id: String;
    patient_id: String;
    procedure_code: String;
    description: String;
    pos: String;
    unit: String;
    procedure_date: String;
    default_amount: String;
    default_modifier: String;
    default_ndc: String;
    practice_id: String;
    system_ip: String;
    deleted: Boolean;
    created_user: String;
    client_date_created: String;
    modified_user: String;
    client_date_modified: String;
    date_created: String;
    date_modified: String;
    comments: String;
    education_provided: Boolean;
    code_type: String;
    entry_type: String;
    from_followup: Boolean;
}